Refactor TestEntityManager so that its not a context and has a proper namespace so that we can use it without registering with the behat.
